change makefile with lib

+++ controller/inc/utils.hpp
@@ -22,7 +22,7 @@ bool Str2Unsint(const string& str, unsigned& unsint) {
     unsint = 0;
     if (str.size() == 0)
         return false;
-    for (int i = 0; i < str.size(); ++i) {
+    for (size_t i = 0; i < str.size(); ++i) {
         if (!isdigit(str[i]))
             return false;
         unsint = unsint*10 + (str[i]-'0');
